intermediary-auditor: shipping first version
For a given list of hostnames, this tool dials and starts a TLS
handshake and audits the (raw) Cetificate Chain presented for
inconsistencies between the Issuer Subject Name of the leaf certificate
and the Subject Name of the of the intermediary certificate.

Accepting inputs are, space separated hostnames supplied via stdin or a
path to a csv file provided following the --stats-csv-file flag. The csv
file must contain the hostname in the first column of every row and must
be in the following format <tld label> followed by each <label> of the
hostname in reverse order (i.e. mail.google.com would be
com.google.mail)

Provided also is a debug mode. If the operator supplies the flag --debug
all hostnames audited will have their results printed in the following
format: leafCert: [subjectCN: <subjectCN> | issuerCN: <issuerCN>] ->
chainCert0: [subjectCN: <subjectCN> | issuerCN: <issuerCN>]