chain-auditor: shipping first version (#26)
For a given list of hostnames, this tool dials and starts a TLS
handshake and audits the presented (raw) certificate chain for
inconsistencies between the Issuer Subject Name of the leaf certificate
and the Subject Name of the of the intermediate certificate.

Acceptable inputs are, space separated hostnames supplied as arguments
or a path to a TSV (tab separated value) file provided following the
--stats-tsv-file flag. The TSV file must contain the hostname in the
second column of every row and the hostname must be in the following
format `<tld label>` followed by each `<label>` of the hostname in
reverse order (i.e. mail.google.com would be com.google.mail)

Provided also is a debug mode. If the operator supplies the flag --debug
all hostnames audited will have their results printed in the following
format: `leafCert: [subjectCN: <subjectCN> | issuerCN: <issuerCN>] ->
chainCert0: [subjectCN: <subjectCN> | issuerCN: <issuerCN>]`

Unrelated: Added the .golangci.yml with lints config from boulder.