Feature Update to Chain Auditor (#28)
Outputs to a TSV file instead of stdout. Added a flag to adjust parallelism.
Hostnames are now randomly shuffled, reducing the risk that we inundate the same
subscriber or partner with requests. Added a progress bar that tracks: percent
complete, rate of completion, remaining count, and errors by count. Refactored
the result returned to a struct containing the info needed for the progress bar
and (later) Prometheus stats. Added all of the logic to populate everything but
the agent version in the result struct. Added a filter for hostnames containing
*, as those will never resolve; this one change reduces the number of hostnames
for a given day by ~10% Removed ChainToString.